+++ test/index.js
@@ -6,6 +6,8 @@ var time = require('unix-time');
 var should = require('should');
 var assert = require('assert');
 var Iterable = require('..');
+var objCase = require('obj-case');
+var mapper = require('../lib/mapper');
 
 describe('Iterable', function(){
   var settings;
@@ -15,9 +17,10 @@ describe('Iterable', function(){
 
   beforeEach(function(){
     payload = {};
-    settings = { apiKey: 'a1b0ad8c09cb419498ab90c2b005ed6a' };
+    settings = { apiKey: '124175f3654446babf5b5966e232d91d' };
     iterable = new Iterable(settings);
     test = Test(iterable, __dirname);
+    test.mapper(mapper);
   });
 
   it('should have correct settings', function(){
@@ -25,8 +28,7 @@ describe('Iterable', function(){
       .name('Iterable')
       .endpoint('https://api.iterable.com/api')
       .channels(['server', 'mobile', 'client'])
-      .ensure('settings.apiKey')
-      .ensure('message.userId');
+      .ensure('settings.apiKey');
   });
 
   describe('.validate()', function(){
@@ -46,12 +48,23 @@ describe('Iterable', function(){
       test.invalid(msg, settings);
     });
 
-    it('should be invalid when .userId is missing', function(){
+    it('should be valid when .apiKey is given', function(){
+      test.valid(msg, settings);
+    });
+
+    it('should be invalid when both .userId and .email are missing', function(){
       delete msg.userId;
+      delete msg.properties.email;
       test.invalid(msg, settings);
     });
 
-    it('should be valid when .apiKey is given', function(){
+    it('should be valid when .userId is missing but .email is present', function(){
+      delete msg.userId;
+      test.valid(msg, settings);
+    });
+
+    it('should be valid when .userId is present but .email is missing', function(){
+      delete msg.properties.email;
       test.valid(msg, settings);
     });
   });
@@ -61,21 +74,55 @@ describe('Iterable', function(){
       it('should map basic track', function(){
         test.maps('track-basic');
       });
+
+      it('should map track with email but no userId', function(){
+        test.maps('track-email-no-userId');
+      });
+
+      it('should map track with userId but no email', function(){
+        test.maps('track-userId-no-email');
+      });
+
+      it('should map added product track', function(){
+        test.maps('track-added-product');
+      });
+
+      it('should map removed product track', function(){
+        test.maps('track-removed-product');
+      });
+
+      it('should map purchase track', function(){
+        test.maps('track-purchase');
+      });
     });
 
     describe('identify', function(){
       it('should map basic identify', function(){
         test.maps('identify-basic');
       });
+
+      it('should map a more advanced identify', function(){
+        test.maps('identify-advanced');
+      });
+
+      it('should map identify with email but no userId', function(){
+        test.maps('identify-email-no-userId');
+      });
+
+      it('should map identify with userId but no email', function(){
+        test.maps('identify-userId-no-email');
+      });
     });
   });
 
   describe('.track()', function(){
-    it('should get a good response from the API', function (done) {
+    it('should get a good response from the API', function(done){
+      var json = test.fixture('track-basic');
       test
         .set(settings)
-        .track(helpers.track())
+        .track(json.input)
         .expects(200)
+        .pathname('/api/events/track')
         .end(done);
     });
 
@@ -85,10 +132,84 @@ describe('Iterable', function(){
         .track({})
         .error('cannot POST /api/events/track (401)', done);
     });
+
+    it('should map Added Product to updateCart if there is a cart provided', function(done){
+      var json = test.fixture('track-added-product');
+
+      test
+        .set(settings)
+        .track(json.input)
+        .sends(json.output)
+        .expects(200)
+        .pathname('/api/commerce/updateCart')
+        .end(done);
+    });
+
+    it('should not map Added Product to updateCart if there is no cart provided', function(done){
+      var json = test.fixture('track-added-product');
+      var input = json.input;
+      input = objCase.del(input, 'properties.products');
+
+      test
+        .set(settings)
+        .track(input)
+        .expects(200)
+        .pathname('/api/events/track')
+        .end(done);
+    });
+
+    it('should map Removed Product to updateCart if there is a cart provided', function(done){
+      var json = test.fixture('track-removed-product');
+
+      test
+        .set(settings)
+        .track(json.input)
+        .sends(json.output)
+        .expects(200)
+        .pathname('/api/commerce/updateCart')
+        .end(done);
+    });
+
+    it('should not map Removed Product to updateCart if there is no cart provided', function(done){
+      var json = test.fixture('track-removed-product');
+      var input = json.input;
+      input = objCase.del(input, 'properties.products');
+
+      test
+        .set(settings)
+        .track(input)
+        .expects(200)
+        .pathname('/api/events/track')
+        .end(done);
+    });
+
+    it('should map Completed Order to trackPurchase', function(done){
+      var json = test.fixture('track-purchase');
+
+      test
+        .set(settings)
+        .track(json.input)
+        .sends(json.output)
+        .expects(200)
+        .pathname('/api/commerce/trackPurchase')
+        .end(done);
+    });
+
+    it('should map a more complex Completed Order to trackPurchase', function(done){
+      var json = test.fixture('track-purchase2');
+
+      test
+        .set(settings)
+        .track(json.input)
+        .sends(json.output)
+        .expects(200)
+        .pathname('/api/commerce/trackPurchase')
+        .end(done);
+    });
   });
 
   describe('.identify()', function(){
-    it('should get a good response from the API', function (done) {
+    it('should get a good response from the API', function(done){
       test
         .set(settings)
         .identify(helpers.identify())
@@ -105,7 +226,7 @@ describe('Iterable', function(){
   });
 
   describe('.page()', function(){
-    it('should not track automatically', function (done) {
+    it('should not track automatically', function(done){
       test
         .set(settings)
         .page(helpers.page())
@@ -113,7 +234,7 @@ describe('Iterable', function(){
         .end(done);
     });
 
-    it('should track all pages', function (done) {
+    it('should track all pages', function(done){
       test
         .set(settings)
         .set({ trackAllPages: true })
@@ -123,7 +244,7 @@ describe('Iterable', function(){
         .end(done);
     });
 
-    it('should track named pages', function (done) {
+    it('should track named pages', function(done){
       test
         .set(settings)
         .set({ trackNamedPages: true })
@@ -133,7 +254,7 @@ describe('Iterable', function(){
         .end(done);
     });
 
-    it('should track category pages', function (done) {
+    it('should track category pages', function(done){
       test
         .set(settings)
         .set({ trackCategorizedPages: true })