+++ History.md
@@ -1,8 +1,23 @@
 
-1.1.1 / 2016-02-16
+1.2.3 / 2016-02-16
 ==================
 
- * allow calls as long as `userId` or `email`
+  * allow calls as long as `userId` or `email`
+
+1.2.2 / 2015-09-02
+==================
+
+  * remove revenue->amount alias per iterable request
+
+1.2.1 / 2015-08-27
+==================
+
+  * still send call to updateCart api when products is an empty array
+
+1.2.0 / 2015-08-26
+==================
+
+  * map ecommerce calls to iterable commerce api
 
 1.1.0 / 2015-04-03
 ==================